real

+++ src/Discord.Net.WebSocket/Entities/AuditLogs/AuditLogCache.cs
@@ -19,7 +19,7 @@ internal class AuditLogCache
     {
         _size = client.AuditLogCacheSize;
         var dictSize = _size;
-        if ((long)dictSize * 1.05 > int.MaxValue)
+        if (dictSize * 1.05 > int.MaxValue)
             dictSize = int.MaxValue;
         else
             dictSize = (int)(dictSize * 1.05);
